+++ src/plone/rest/service.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
-from Products.Five import BrowserView
 from plone.rest.interfaces import IService
 from zope.interface import implements
 
 import json
 
 
-class Service(BrowserView):
+class Service(object):
     implements(IService)
 
     def __call__(self):